+++ pkg/sequence/lexer/constructors.go
@@ -7,6 +7,34 @@ import (
 	"github.com/ufukty/diagramer/pkg/sequence/lexer/tokens"
 )
 
+// selective line constructor
+type constructor interface {
+	check(string) bool
+	construct(string) Line
+}
+
+// internal purposes only
+var precedence = []constructor{
+	&Activate{},
+	&Alt{},
+	&And{},
+	&Box{},
+	&Break{},
+	&Create{},
+	&Critical{},
+	&Deactivate{},
+	&Destroy{},
+	&Else{},
+	&End{},
+	&LifelineDecl{},
+	&Loop{},
+	&Message{},
+	&Note{},
+	&Option{},
+	&Parallel{},
+	&WideNote{},
+}
+
 var (
 	regexActivate   = regexp.MustCompile(`activate\s+(\w+)`)
 	regexAlt        = regexp.MustCompile(`alt\s+(.+)`)
